nothing

+++ mandelbrot/numpy_vector/numpy_vector.py
@@ -1,7 +1,6 @@
 import datetime
 import sys
 import numpy as nm
-import Image
 
 # area of space to investigate
 x1, x2, y1, y2 = -2.13, 0.77, -1.3, 1.3
@@ -16,6 +15,7 @@ def calculate_z_numpy(q, maxiter, z):
         z = nm.where(done,0+0j, z)
         output = nm.where(done, iteration, output)
     return output
+
 
 def calculate(show_output):
     # make a list of x and y values
@@ -32,13 +32,13 @@ def calculate(show_output):
 
     print "xx and yy have length", len(xx), len(yy)
 
-    # yy will become 0+yyj when cast to complex64 (2 * 4byte float32) 
-    yy = yy.astype(nm.complex64)
+    # yy will become 0+yyj when cast to complex128 (2 * 8byte float64) same as Python float 
+    yy = yy.astype(nm.complex128)
     # create q as a square matrix initially of complex numbers we're calculating
     # against, then flatten the array to a vector
-    q = nm.ravel(xx+yy[:, nm.newaxis]).astype(nm.complex64)
+    q = nm.ravel(xx+yy[:, nm.newaxis]).astype(nm.complex128)
     # create z as a 0+0j array of the same length as q
-    z = nm.zeros(q.shape, nm.complex64)
+    z = nm.zeros(q.shape, nm.complex128)
 
     start_time = datetime.datetime.now()
     print "Total elements:", len(q)
@@ -51,6 +51,7 @@ def calculate(show_output):
     print "Total sum of elements (for validation):", validation_sum
 
     if show_output: 
+        import Image
         output = (output + (256*output) + (256**2)*output) * 8
         im = Image.new("RGB", (w/2, h/2))
         im.fromstring(output.tostring(), "raw", "RGBX", 0, -1)
